requirements.txt
# requirements.txt
pytesseract
pandas
Pillow
opencv-python